move PrimTrail to PrimTrail
